upharma-au/requirements.txt
httpx
selectolax
trafilatura
supabase
python-dotenv
tenacity
openai